ebtookdak/requirements.txt
APScheduler==3.9.1.post1
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
djangorestframework==3.14.0
idna==3.4
octorest==0.4
pytz==2022.6
<<<<<<< HEAD
=======
pytz-deprecation-shim==0.1.0.post0
>>>>>>> 63ee125122480c38ec679cec65371872eae21fa5
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.12
websocket-client==1.4.1
psycopg2==2.8.4
